Update text_parser to handle name and street name stops
Using City of Chicago Open Data portal and a list pre-collected by an
organization in the space (i.e. not determined from NER), create
name_stops.py, a file with constants for a long list of names that
could be in the dataset, a list of every street name in the City, and
a list of every street suffix (e.g.: st, ave).

Also, add write_to_file option for get_text_corpus() method, so that you can
save a corpus and reload it later if needed. Note that a corpus is usually between
30 and 50 MB, so make sure you have space to store it. Defaults to using a
../../corpora folder (i.e. directly under repository root directory).

Removing the longer list of stopwords increases runtime of get_text_corpus()
from 10-30 seconds to 10-15 minutes. Lemmatizing also slightly increases runtime.